Make use of newly added instructions
With 3089ef6, the Go project finally supports VUSRA, FMOVQ, VUADDW, and
VUADDW2.  Make use of them.  Unfortunately, not all addressing modes
seem to be supported for FMOVQ and support for VUADDL and VUADDL2
has not yet landed.  This will be added in a future commit.

See also https://github.com/golang/go/issues/40725