countavx512_amd64.s: greatly simplify tail handling
It occurred to me that the tail handling code is essentially
computing a mask of the bits in each byte in the mask registers.
Instead of computing the masks, we can simply move the data to
the mask registers directly, simplifying the code.

The new runtrunt case handles very short inputs for which we
cannot cheat by overreading here and there.

This yields an overall 5% speedup over the previous approach,
with up to 25% speedup for the 10 byte input case.