countsse2_amd64.s: fix overreads
Turns out MOVD assembles to movq instead of movd as expected.
To get movd, the MOVL mnemonic is needed.  This went unnoticed
until the overreading tests were added.

Also fix an off-by-one error in the runt case, where a direct
load from the buffer was attempted even for cases where the
buffer did in fact not cross the end end of memory.